+++ reactnative/src/scenes/Fridge.js
@@ -55,11 +55,11 @@ class Fridge extends Component {
         return (
             <Container style={styles.container}>
                 <Header
-                    style={{ backgroundColor: '#5BB85D' }}
+                    style={{ backgroundColor: '#f2f2f2' }}
                     androidStatusBarColor="#000"
                 >
                     <Body>
-                        <Title style={{ color: '#fff' }}>My Fridge</Title>
+                        <Title style={{ color: '#000' }}>My Fridge</Title>
                     </Body>
                 </Header>
                 <Content style={{ flex: 1, backgroundColor: '#fff', marginTop: 5 }}>
